expanded all "inner" classes
