+++ src/main/kotlin/dev/wisplang/wisp/Primitive.kt
@@ -1,17 +1,19 @@
 package dev.wisplang.wisp
 
+import dev.wisplang.wisp.tokenizer.Tokenizer
+
 enum class Primitive( type: BaseType, size: Short ) {
-    U1( BaseType.Unsigned, 1 ),
-    U8( BaseType.Unsigned, 8 ),
-    I8( BaseType.Integer, 8 ),
-    U16( BaseType.Unsigned, 16 ),
-    I16( BaseType.Integer, 16 ),
-    U32( BaseType.Unsigned, 32 ),
-    I32( BaseType.Integer, 32 ),
-    F32( BaseType.Float, 32 ),
-    U64( BaseType.Unsigned, 64 ),
-    I64( BaseType.Integer, 64 ),
-    F64( BaseType.Float, 64 );
+    U1( BaseType.UNSIGNED, 1 ),
+    U8( BaseType.UNSIGNED, 8 ),
+    I8( BaseType.INTEGER, 8 ),
+    U16( BaseType.UNSIGNED, 16 ),
+    I16( BaseType.INTEGER, 16 ),
+    U32( BaseType.UNSIGNED, 32 ),
+    I32( BaseType.INTEGER, 32 ),
+    F32( BaseType.FLOAT, 32 ),
+    U64( BaseType.UNSIGNED, 64 ),
+    I64( BaseType.INTEGER, 64 ),
+    F64( BaseType.FLOAT, 64 );
 
     override fun toString() = this.name.lowercase()
 
@@ -22,8 +24,8 @@ enum class Primitive( type: BaseType, size: Short ) {
     }
 
     enum class BaseType( code: String ) {
-        Unsigned( "u" ),
-        Integer( "i" ),
-        Float( "f" )
+        UNSIGNED( "u" ),
+        INTEGER( "i" ),
+        FLOAT( "f" )
     }
 }
