+++ src/main/kotlin/dev/wisplang/wisp/ast/DefinedFunction.kt
@@ -1,0 +1,8 @@
+package dev.wisplang.wisp.ast
+
+data class DefinedFunction(
+    val name: String,
+    val returnType: BaseType,
+    val parameters: Map<String, BaseType>,
+    val operators: Unit
+)
