+++ src/main/kotlin/dev/wisplang/wisp/ast/Root.kt
@@ -1,0 +1,7 @@
+package dev.wisplang.wisp.ast
+
+data class Root(
+    val types: Map<String, DefinedType>,
+    val globals: Map<String, DefinedVariable>,
+    val functions: Map<String, DefinedFunction>
+)
