+++ src/main/kotlin/dev/wisplang/wisp/tokenizer/MatureToken.kt
@@ -1,0 +1,3 @@
+package dev.wisplang.wisp.tokenizer
+
+data class MatureToken( val type: MatureType, val value: String, val idx: Int, val len: Int )
