+++ src/main/kotlin/dev/wisplang/wisp/tokenizer/MatureType.kt
@@ -1,0 +1,12 @@
+package dev.wisplang.wisp.tokenizer
+
+enum class MatureType {
+    SYMBOL,
+    KEYWORD,
+    INTEGER,
+    FLOAT,
+    STRING,
+    NAME,
+    NEWLINE,
+    PRIMITIVE
+}
