+++ src/main/kotlin/dev/wisplang/wisp/tokenizer/Token.kt
@@ -1,0 +1,3 @@
+package dev.wisplang.wisp.tokenizer
+
+data class Token( val type: Type, val value: String, val idx: Int, val len: Int )
