+++ src/main/kotlin/dev/wisplang/wisp/tokenizer/Type.kt
@@ -1,0 +1,9 @@
+package dev.wisplang.wisp.tokenizer
+
+enum class Type {
+    SYMBOL,
+    NUMBER,
+    STRING,
+    NAME,
+    NEWLINE
+}
